Move the Plane in v_frame
