+++ benches/rdo.rs
@@ -11,7 +11,7 @@ use rav1e::prelude::Plane;
 fn init_plane_u8(width: usize, height: usize, seed: u8) -> Plane<u8> {
   let mut ra = ChaChaRng::from_seed([seed; 32]);
   let data: Vec<u8> = (0..(width * height)).map(|_| ra.gen()).collect();
-  Plane::wrap(data, width)
+  Plane::from_slice(&data, width)
 }
 
 pub fn cdef_dist_wxh_8x8(c: &mut Criterion) {
