+++ src/asm/x86/predict.rs
@@ -305,7 +305,7 @@ mod test {
       };
       for angle in angles {
         let expected = {
-          let mut plane = Plane::wrap(vec![0u8; 4 * 4], 4);
+          let mut plane = Plane::from_slice(&vec![0u8; 4 * 4], 4);
           native::dispatch_predict_intra(
             *mode,
             *variant,
@@ -325,7 +325,7 @@ mod test {
           data
         };
 
-        let mut output = Plane::wrap(vec![0u8; 4 * 4], 4);
+        let mut output = Plane::from_slice(&vec![0u8; 4 * 4], 4);
         dispatch_predict_intra(
           *mode,
           *variant,
